+++ cmd/seid/cmd/root.go
@@ -31,7 +31,7 @@ import (
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	"github.com/cosmos/cosmos-sdk/x/crisis"
 	genutilcli "github.com/cosmos/cosmos-sdk/x/genutil/client/cli"
-	xwasm "github.com/sei-protocol/sei-chain/wasm"
+	"github.com/sei-protocol/sei-chain/wasmbinding"
 	"github.com/spf13/cast"
 	"github.com/spf13/cobra"
 	tmcli "github.com/tendermint/tendermint/libs/cli"
@@ -242,7 +242,7 @@ func newApp(
 	}
 
 	wasmopts := []wasm.Option{wasmkeeper.WithMessageEncoders(&wasmkeeper.MessageEncoders{
-		Custom: xwasm.DexCustomEncoder,
+		Custom: wasmbinding.CustomEncoder,
 	})}
 
 	return app.New(
